add module structure for data analyze(data_nalyzer.py); script management (import_data.py); and Makefile commands refs #33430
